API break: generate_json_result takes a list as input, instead of a str
The list can contains str, or any elements that can be converted to a str with source mapping
Ex: Variable, Contract, Event, Function, ..
generate_json_result will automatically add the elements to the json.
As a result, detectors only need to add the elements to the info list, and do not need to add the element
to the json.

AbstractDetectors has a STANDARD_JSON parameter, if set to False, generate_json_result will not add the elements to the json