Handle ignore markers when `exclude-dependencies` is set
`valid_results` behaviour is described in terms of "results are valid,
unless ...".  A matching implementation would be to only return `True` a
the end of the function, and have all the cases only return `False` if
they detect an invalid result.

`show-ignored-findings` is the only exception to the rule above.