GitHub Super Linter: ignore test json, zip, sol files
These test files are a large part of the repository, and do not need to be
linted. The large number of files makes the enumeration slow when performed
by the Super Linter, which results in CI jobs taking considerably long.

This PR ignores said files in hopes to speed up the CI workflows.
It also switches the Super Linter image to the slim version, which should
be slightly faster to download.

Fixes: #1084